[REFACTOR] generalize mark component
